+++ lib/src/esc_pos_base.dart
@@ -47,6 +47,8 @@ class PrinterDocument {
       addCommand(PrinterCommandImage(image, align: align));
 
   void print(GenericPrinter printer) {
+    printer.reset();
+
     for (var c in commands) {
       c.print(printer);
     }